refactor: update audio handling and constants for favorites and popout grids
